Spell Lookup Table
The spell lookup table is a table where each word in the spell names maps to a list of spell data containing that word. This table is used to quickly find all spells that contain a given word, making the tooltip text checking a lot faster.
Also, added sounds to Toggling Inspect Mode.